requirements.txt
Django==1.7.1
Pillow==2.6.1
South==1.0.1
Unidecode==0.04.16
argparse==1.2.2
cmsplugin-filer==0.10.1
dj-database-url==0.3.0
django-appconf==0.6
django-classy-tags==0.5.2
django-cms==3.0.7
django-compressor==1.4
django-filer==0.9.8
django-mptt==0.6.1
django-ordered-model==0.4.0
django-polymorphic==0.6
django-reversion==1.8.5
django-sekizai==0.7
djangocms-admin-style==0.2.4
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.6.0
djangocms-link==1.5
djangocms-picture==0.1
djangocms-snippet==1.3
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.1
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.999
python-slugify==0.1.0
pytz==2014.10
requests==2.5.0
six==1.8.0
wsgiref==0.1.2
